Add compatibility with doctrine/collections v3, add SortDirection support in Query Builder (#3014)
