+++ composer.json
@@ -28,7 +28,7 @@
     "require": {
         "php": "^8.1",
         "ext-mongodb": "^1.21 || ^2.0",
-        "doctrine/collections": "^2.1",
+        "doctrine/collections": "^2.1 || ^3.1",
         "doctrine/event-manager": "^1.0 || ^2.0",
         "doctrine/instantiator": "^1.1 || ^2",
         "doctrine/persistence": "^3.2 || ^4",
@@ -39,6 +39,7 @@
         "symfony/console": "^5.4 || ^6.4 || ^7.0 || ^8.0",
         "symfony/deprecation-contracts": "^2.2 || ^3.0",
         "symfony/polyfill-php84": "^1.33",
+        "symfony/polyfill-php86": "^1.37",
         "symfony/var-dumper": "^5.4 || ^6.4 || ^7.0 || ^8.0",
         "symfony/var-exporter": "^6.4.1 || ^7.0 || ^8.0"
     },
